v1.3.4

+++ SteamAutoMarketWPF/SteamAutoMarket/SteamAutoMarket/Steam/Market/Interface/MarketClient.cs
@@ -829,7 +829,7 @@
 
         public dynamic SellItem(int appId, int contextId, long assetId, int amount, double priceWithFee)
         {
-            var priceWithoutFee = int.Parse(this.GetSteamPriceWithoutFee(priceWithFee));
+            var priceWithoutFee = double.Parse(this.GetSteamPriceWithoutFee(priceWithFee));
             var data = new Dictionary<string, string>
                            {
                                { "appid", appId.ToString() },
@@ -837,7 +837,7 @@
                                { "contextid", contextId.ToString() },
                                { "assetid", assetId.ToString() },
                                { "amount", amount.ToString() },
-                               { "price", (priceWithoutFee * 100).ToString() }
+                               { "price", ((int)(priceWithoutFee * 100)).ToString() }
                            };
 
             var resp = this.steam.Request(Urls.Market + "/sellitem/", Method.POST, Urls.Market, data, true).Data
